inventario_avanzado3/requirements.txt
django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
dash==2.17.1
plotly==5.24.1
